Refactor api abstraction.
